+++ include/exiv2/metadatum.hpp
@@ -219,9 +219,9 @@ namespace Exiv2 {
         //! Return the size in bytes of one component of this type
         virtual size_t typeSize() const =0;
         //! Return the number of components in the value
-        virtual long count() const =0;
+        virtual size_t count() const =0;
         //! Return the size of the value in bytes
-        virtual long size() const =0;
+        virtual size_t size() const =0;
         //! Return the value as a string.
         virtual std::string toString() const =0;
         /*!
